requirements.txt
aiogram==3.13.1
aiohttp
aiodns
pyjwt[crypto]
bs4